Moved tests to original file
