+++ pandas/tests/strings/test_find_replace.py
@@ -328,6 +328,44 @@ def test_contains_compiled_regex_flags(any_string_dtype):
     expected = Series([False, True, True], dtype=expected_dtype)
     tm.assert_series_equal(result, expected)
 
+@pytest.mark.parametrize(
+        "values, pat, expected_values",
+        [
+            (
+                ["hello hello", "world world", "foo bar"],
+                r"(\w+)\s\1",
+                [True, True, False],
+            ),
+            (
+                ["aa bb", "cc dd", "ab cd"],
+                r"((\w)\2)\s((\w)\4)",
+                [True, True, False],
+            ),
+            (
+                ["hello hello", "foo bar"],
+                r"(?P<word>\w+)\s(?P=word)",
+                [True, False],
+            ),
+        ],
+    )
+def test_contains_backreferences(
+        nullable_string_dtype, values, pat, expected_values
+    ):
+    import warnings
+
+    ser = Series(values, dtype=nullable_string_dtype)
+
+    with warnings.catch_warnings():
+        warnings.filterwarnings(
+            "ignore",
+            message="This pattern is interpreted as a regular expression",
+            category=UserWarning,
+        )
+        result = ser.str.contains(pat, regex=True)
+
+    expected = Series(expected_values, dtype=object)
+
+    tm.assert_series_equal(result, expected, check_dtype=False)
 
 # --------------------------------------------------------------------------------------
 # str.startswith
