+++ pandas/tests/strings/test_string_array.py
@@ -124,50 +124,3 @@ def test_string_array_zfill(nullable_string_dtype, values, width, expected):
     result = s.str.zfill(width)
     expected = Series(expected, dtype=nullable_string_dtype)
     tm.assert_series_equal(result, expected)
-
-
-class TestGH63739:
-    """
-    Test that backreferences trigger fallback from RE2 to Python regex.
-    GH#63739
-    """
-
-    @pytest.mark.parametrize(
-        "values, pat, expected_values",
-        [
-            (
-                ["hello hello", "world world", "foo bar"],
-                r"(\w+)\s\1",
-                [True, True, False],
-            ),
-            (
-                ["aa bb", "cc dd", "ab cd"],
-                r"((\w)\2)\s((\w)\4)",
-                [True, True, False],
-            ),
-            (
-                ["hello hello", "foo bar"],
-                r"(?P<word>\w+)\s(?P=word)",
-                [True, False],
-            ),
-        ],
-    )
-    def test_contains_backreferences(
-        self, nullable_string_dtype, values, pat, expected_values
-    ):
-        import warnings
-
-        ser = Series(values, dtype=nullable_string_dtype)
-
-        # Ignore warning about capture groups (required for backrefs)
-        with warnings.catch_warnings():
-            warnings.filterwarnings(
-                "ignore",
-                message="This pattern is interpreted as a regular expression",
-                category=UserWarning,
-            )
-            result = ser.str.contains(pat, regex=True)
-
-        expected = Series(expected_values, dtype=object)
-
-        tm.assert_series_equal(result, expected, check_dtype=False)
